+++ examples/src/todomvc/src/livestore/mutations.ts
@@ -1,48 +1,45 @@
-import { defineMutation, sql } from '@livestore/livestore'
+import { defineMutation } from '@livestore/livestore'
 import { Schema } from 'effect'
 
 import { Filter } from '../types.js'
+import { tables } from './schema.js'
 
 export const todoCreated = defineMutation(
   'todoCreated',
   Schema.Struct({ id: Schema.String, text: Schema.String }),
-  sql`INSERT INTO todos (id, text, completed) VALUES ($id, $text, false)`,
+  ({ id, text }) => tables.todos.query.insert({ id, text }),
 )
 
-export const todoCompleted = defineMutation(
-  'todoCompleted',
-  Schema.Struct({ id: Schema.String }),
-  sql`UPDATE todos SET completed = true WHERE id = $id`,
+export const todoCompleted = defineMutation('todoCompleted', Schema.Struct({ id: Schema.String }), ({ id }) =>
+  tables.todos.query.update({ completed: true }).where({ id }),
 )
 
-export const todoUncompleted = defineMutation(
-  'todoUncompleted',
-  Schema.Struct({ id: Schema.String }),
-  sql`UPDATE todos SET completed = false WHERE id = $id`,
+export const todoUncompleted = defineMutation('todoUncompleted', Schema.Struct({ id: Schema.String }), ({ id }) =>
+  tables.todos.query.update({ completed: false }).where({ id }),
 )
 
 export const todoDeleted = defineMutation(
   'todoDeleted',
   Schema.Struct({ id: Schema.String, deleted: Schema.DateFromNumber }),
-  sql`UPDATE todos SET deleted = $deleted WHERE id = $id`,
+  ({ id, deleted }) => tables.todos.query.update({ deleted }).where({ id }),
 )
 
 export const todoClearedCompleted = defineMutation(
   'todoClearedCompleted',
-  Schema.Struct({ deleted: Schema.Number }),
-  sql`UPDATE todos SET deleted = $deleted WHERE completed = true`,
+  Schema.Struct({ deleted: Schema.DateFromNumber }),
+  ({ deleted }) => tables.todos.query.update({ deleted }).where({ completed: true }),
 )
 
 export const updatedNewTodoText = defineMutation(
   'updatedNewTodoText',
   Schema.Struct({ text: Schema.String, sessionId: Schema.String }),
-  sql`UPDATE app SET newTodoText = $text WHERE id = $sessionId`,
+  ({ text, sessionId }) => tables.app.query.update({ newTodoText: text }).where({ id: sessionId }),
   { clientOnly: true },
 )
 
 export const filterUpdated = defineMutation(
   'setFilter',
   Schema.Struct({ filter: Filter, sessionId: Schema.String }),
-  sql`UPDATE app SET filter = $filter WHERE id = $sessionId`,
+  ({ filter, sessionId }) => tables.app.query.update({ filter }).where({ id: sessionId }),
   { clientOnly: true },
 )